requirements.txt
fastapi~=0.78.0
uvicorn~=0.16.0
pydantic~=1.9.0
python-dotenv~=0.19.0
requests~=2.27.1

# uvicorn[standard] dependencies
#uvloop~=0.16.0
httptools~=0.3.0
websockets~=10.1
PyYAML~=6.0
CairoSVG~=2.5.2